layout/base/tests/chrome/test_get_printer_paper_sizes.html

<!DOCTYPE HTML>
<html>
<head>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body onload="run()">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

async function run() {
  try {
    let printerList = Cc["@mozilla.org/gfx/printerlist;1"].getService(
      Ci.nsIPrinterList
    );
    if (printerList.printers.length == 0) {
      ok(true, "There were no printers to iterate through.");
    }

    for (let printer of printerList.printers) {
      isnot(printer.name, null, "Printer name should never be null.");
      isnot(printer.name, "", "Printer name should never be empty.");
      info(printer.name);

      let paperList = await printer.paperList;
      for (let paper of paperList) {
        paper.QueryInterface(Ci.nsIPaper);

        info(`${paper.name}: ${paper.width}x${paper.height} with margin: ${paper.unwriteableMarginTop} ${paper.unwriteableMarginRight} ${paper.unwriteableMarginBottom} ${paper.unwriteableMarginLeft}`);
        isnot(paper.name, null, "Paper name should never be null.");
        isnot(paper.name, "", "Paper name should never be empty.");

        isnot(paper.width, null, "Paper width should never be null.");
        ok(paper.width > 0.0, "Paper width should be greater than zero.");

        isnot(paper.height, null, "Paper height should never be null.");
        ok(paper.height > 0.0, "Paper height should be greater than zero.");

        isnot(paper.unwriteableMarginTop, null, "Paper unwriteableMarginTop should never be null.");
        ok(paper.unwriteableMarginTop >= 0.0, "Paper unwriteableMarginTop should be greater than or equal to zero.");

        isnot(paper.unwriteableMarginBottom, null, "Paper unwriteableMarginBottom should never be null.");
        ok(paper.unwriteableMarginBottom >= 0.0, "Paper unwriteableMarginBottom should be greater than or equal to zero.");

        isnot(paper.unwriteableMarginLeft, null, "Paper unwriteableMarginLeft should never be null.");
        ok(paper.unwriteableMarginLeft >= 0.0, "Paper unwriteableMarginLeft should be greater than or equal to zero.");

        isnot(paper.unwriteableMarginRight, null, "Paper unwriteableMarginRight should never be null.");
        ok(paper.unwriteableMarginRight >= 0.0, "Paper unwriteableMarginRight should be greater than or equal to zero.");
      }
    }
  } catch (e) {
    Cu.reportError(e);
  }
  SimpleTest.finish();
}

</script>
</body>
</html>
